Added Mouse and enchanced classes
I added a mouse class (that changes the cursor to a hand, might be nice
to allow players to choose their options in the menu with a nice
cursor). Also, I have added LivingObject and DeadObject classes that
refer to objects that move constantly and take part in the gameplay and
those that don't. Futhermore the collision between magefireballs and map
fireballs has been implemented in O(n*m) which is n for the first team
objects and m for the second team objects.